# api/openapi/api.yaml
openapi: 3.0.1
info:
  title: Restoo API
  description: API document for Restoo app
  termsOfService: 'http://swagger.io/terms/'
  version: 1.0.0
...
servers:
  - url: 'http://localhost:3001/api/'
  - url: 'https://restoo-backend.herokuapp.com/api/'
...
paths:
...
  /order/list:
    get:
      description: List all orders
      summary: List all orders
      security:
        - bearerAuth: []
      tags:
        - Waiter
        - Admin
        - Order
      operationId: listOrders
      parameters:
        - $ref: '#/components/parameters/LimitQueryParam'
        - $ref: '#/components/parameters/OffsetQueryParam'
        - $ref: '#/components/parameters/OrderStatusQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
        - $ref: '#/components/parameters/UserIdQueryParam'
        - $ref: '#/components/parameters/WaiterIdQueryParam'
      responses:
        '200':
          $ref: '#/components/responses/OrderLisitingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/UnknownError'
  /order/list/my:
    get:
      description: List my orders
      summary: List my orders
      security:
        - bearerAuth: []
      tags:
        - User
        - Order
      operationId: listMyOrders
      parameters:
        - $ref: '#/components/parameters/LimitQueryParam'
        - $ref: '#/components/parameters/OffsetQueryParam'
        - $ref: '#/components/parameters/OrderStatusQueryParam'
        - $ref: '#/components/parameters/SortQueryParam'
      responses:
        '200':
          $ref: '#/components/responses/OrderLisitingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/UnknownError'
...
  '/order/{id}/status':
    put:
      description: Change status of an order
      summary: Change status of an order
      security:
        - bearerAuth: []
      tags:
        - Order
        - Admin
        - Waiter
      operationId: changeOrderStatus
      externalDocs:
        url: 'https://github.com/syetalabs/Restoo_Backend/blob/main/docs/diagrams/Order-State.png'
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatusEnum'
      responses:
        '200':
          $ref: '#/components/responses/OkResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/UnknownError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    OkResponse:
      description: Ok
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    UnknownError:
      description: Unknown Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    ValidationError:
      description: ValidationError Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
...
    OrderLisitingResponse:
      description: order list
      content:
        application/json:
          schema:
            properties:
              total:
                type: integer
                description: total orders
                example: 10
              orders:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/OrderBase'
                  properties:
                    customer:
                      type: string
                      example: 5ff2ae9a4e797a001e87fcdb
...
  parameters:
    IdPathParam:
      in: path
      name: id
      schema:
        type: string
      required: true
    UserIdQueryParam:
      in: query
      name: userId
      schema:
        type: string
      required: false
    WaiterIdQueryParam:
      in: query
      name: waiterId
      schema:
        type: string
      required: false
    OrderStatusQueryParam:
      in: query
      name: status
      schema:
        type: string
        enum:
          - waiting
          - accepted
          - preparing
          - served
          - completed
    OffsetQueryParam:
      in: query
      name: offset
      description: Offset the result
      schema:
        type: integer
        example: 1
    LimitQueryParam:
      in: query
      name: limit
      description: Limit the returned number of results
      schema:
        type: integer
        example: 10
    SortQueryParam:
      in: query
      name: sort
      description: Sort the result
      schema:
        type: string
        enum:
          - asc
          - desc
  schemas:
...
    CreateOrderItemSchema:
      type: object
      title: ''
      description: ''
      properties:
        item:
          type: string
          example: 5ff2b5b81dd33e00ebad1269
          description: item id
        quantity:
          type: integer
          example: 1
          description: quantity of item
        price:
          type: number
          example: 4
        name:
          type: string
          example: cake
...
    OrderStatusEnum:
      type: string
      description: status of order
      enum:
        - waiting
        - accepted
        - preparing
        - served
        - completed
        - cancelled
    OrderBase:
      type: object
      properties:
        _id:
          type: string
          description: Id of order
          example: 5ffd8dc6faa662038750c4a6
        status:
          $ref: '#/components/schemas/OrderStatusEnum'
        note:
          type: string
          description: note
          example: do it quickly
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItemSchema'
        paymentMethod:
          type: string
          enum:
            - cash
            - card
          example: card
        placedAt:
          type: string
          description: order placed at
          example: '2021-01-12T11:53:42.602Z'
        currency:
          type: string
          example: usd
        total:
          type: integer
          format: int32
          description: total price
          example: 120
        table:
          type: integer
          description: table no
          example: 14
        reference:
          type: string
          description: order referenece
          example: 2PrWIV2-_arrAfC_QEf5X
        customer:
          type: object
          properties:
            name:
              type: string
              example: John doe
            email:
              type: string
              example: joe224@gmail.com
            phoneNumber:
              type: string
              example: '0712456787'
...
    MessageResponse:
      type: object
      description: message
      properties:
        message:
          type: string
          example: ok
    APIError:
      type: object
      description: Error object from API
      properties:
        message:
          type: string
          example: invalid token
        extra:
          type: string
          format: nullable
        errors:
          type: object
          properties:
            name:
              type: string
              example: APIError
            message:
              type: string
              example: invalid token
            status:
              type: integer
              format: int32
              example: 401
            extra:
              type: string
              format: nullable
...